Remove todo, rename state_readers file [skip ci]
