+++ crates/starknet-devnet-core/src/state/state_readers.rs
